requirements.txt
asgiref==3.2.3
astroid==2.3.3
colorama==0.4.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.1
django-crispy-forms==1.8.1
django-import-export==2.0
et-xmlfile==1.0.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.0
openpyxl==3.0.2
Pillow==6.2.1
pylint==2.4.4
pytz==2019.3
PyYAML==5.2
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0